# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "aalug",
            "url": "https://github.com/aalug",
            "email": "a.a.gulczynski@gmail.com"
        }
    },
    "paths": {
...
        "/jobs": {
            "get": {
                "description": "Filter and list jobs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Filter and list jobs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Job title - matches partially (ILIKE)",
                        "name": "title",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job industry - exact name",
                        "name": "industry",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job location - exact name",
                        "name": "job_location",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Salary min - must be smaller or equal salary_max",
                        "name": "salary_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Salary max - must be greater or equal salary_min",
                        "name": "salary_max",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/db.ListJobsByFiltersRow"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new job",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Create job",
                "parameters": [
                    {
                        "description": "Job details",
                        "name": "CreateJobRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createJobRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.jobResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/company": {
            "get": {
                "description": "List jobs by company name, id or part of the name.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List jobs by company",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Company ID",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Company name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Part of the company name",
                        "name": "name_contains",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/db.ListJobsByCompanyNameRow"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query. Only one of the three parameters is allowed.",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobs/match-skills": {
            "get": {
                "description": "List jobs that match the authenticated users skills",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "List jobs by matching skills",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/db.ListJobsMatchingUserSkillsRow"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid query",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Employer making the request - only users can access",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Any other error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.createJobRequest": {
            "type": "object",
            "required": [
                "description",
                "industry",
                "location",
                "required_skills",
                "requirements",
                "salary_max",
                "salary_min",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "required_skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "requirements": {
                    "type": "string"
                },
                "salary_max": {
                    "type": "integer",
                    "minimum": 0
                },
                "salary_min": {
                    "type": "integer",
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "api.jobResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "industry": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "required_skills": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.ListJobSkillsByJobIDRow"
                    }
                },
                "requirements": {
                    "type": "string"
                },
                "salary_max": {
                    "type": "integer"
                },
                "salary_min": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "db.ListJobSkillsByJobIDRow": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "skill": {
                    "type": "string"
                }
            }
        },
...
        "db.ListJobsByFiltersRow": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "company_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "requirements": {
                    "type": "string"
                },
                "salary_max": {
                    "type": "integer"
                },
                "salary_min": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}